sem_2/arc/hw_4/requirements.txt
fastapi[standard]
uvicorn[standard]
kafka-python==2.0.2
pydantic
sqlalchemy
asyncpg
psycopg2-binary